Update redirects in _redirects.njk file
The redirection paths in _redirects.njk file for 'aa-updates' have been updated. This will correct the issue of any access requests to 'aa-updates' being redirected to the appropriate place. A new line is also added to keep the URL redirected to the root of the site if 'aa-updates' is accessed directly.